+++ src/components/thumbnails/Cell.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { CardAction, ContextMenu } from 'types/components';
 import Thumbnail from './Thumbnail';
 import { getIndexFromRowCol } from '../../util/utils';
+import { Post } from '../../types/gelbooruTypes';
 
 interface CellProps {
 	columnIndex: number;
@@ -12,6 +13,9 @@ interface CellProps {
 	itemCount: number;
 	contextMenu?: ContextMenu[];
 	actions?: CardAction[];
+	onMouseEnter?: (event: React.MouseEvent<HTMLDivElement, MouseEvent>, post: Post) => void;
+	onMouseLeave?: (event: React.MouseEvent<HTMLDivElement, MouseEvent>, post: Post) => void;
+	onMouseMove?: (event: React.MouseEvent<HTMLDivElement, MouseEvent>, post: Post) => void;
 }
 
 const Cell: React.FunctionComponent<CellProps> = (props) => {
@@ -21,7 +25,15 @@ const Cell: React.FunctionComponent<CellProps> = (props) => {
 	}
 	return (
 		<div style={{ ...props.style, width: '100%' }}>
-			<Thumbnail isScrolling={props.isScrolling} index={index} contextMenu={props.contextMenu} actions={props.actions} />
+			<Thumbnail
+				isScrolling={props.isScrolling}
+				index={index}
+				contextMenu={props.contextMenu}
+				actions={props.actions}
+				onMouseEnter={props.onMouseEnter}
+				onMouseLeave={props.onMouseLeave}
+				onMouseMove={props.onMouseMove}
+			/>
 		</div>
 	);
 };